requirements.txt
python-telegram-bot==21.9
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
reportlab==4.0.4
psycopg2-binary==2.9.9
